Centos Dockerfile upgrade
Adding gluster-setup files which is used on first initialization
of gluster contianer as bind mount of gluster config directory will
be empty. This gluster-setup will copy the initial config files which
is required for glusterd to start.

Adding packages for gluster services to work,
ntp, rsync, tar, openssh-server, openssh-clients : Required for geo-replication
cronie : Required for snapshot-scheduler-script

Signed-off-by: Mohamed Ashiq Liyazudeen <mliyazud@redhat.com>