Fedora: be more efficient for detecting brick mux setting
Previously, we were grep-ing through the output of gluster volume info
for the value of the cluster.brick-multiplex setting. But it is highly
inefficient to use `gluster volume list` to obtain a global setting.
With many volumes, this can take unnecessarily long. Furthermore, since
this is invoked right after the start of glusterd, glusterd may not be
done initializing.

This patch changes the retrieval of the brick multiplex setting
to use the command `gluster volume get all cluster.brick-multiplex`
which is exactly made for this purpose and won't get the whole list
of volumes.

Signed-off-by: Michael Adam <obnox@redhat.com>