Add gluster-rsyslog Alpine container
This container will be mainly expected to be lightweight and run as a
sidecar container in a gcs cluster along with containers running gd2.
This is mainly used to parse gluster specific logs and normalize it for
elastic search. Right now only gd2 logs are parsed to some extent. For
this, an rsyslog configuration along with a rulebase file are included.

Signed-off-by: Sridhar Seshasayee <sseshasa@redhat.com>